refactor: refactored the schedule and device modules
