+++ tests/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Switcher unofficial integration test cases."""
+"""Switcher unofficial integration main module test cases."""
