+++ tests/testresources/test_device_parsing/test_a_power_plug_datagram_produces_device.txt
