+++ tests/testresources/test_device_parsing/test_a_water_heater_datagram_produces_device.txt
